# api.yaml
swagger: "2.0"
info:
  title: TCF API
  description: This is API Documentation of TFC project
  version: 1.0
  contact:
    email: cedricgnangue08@gmail.com
  liscence:
    name: Apache 2.0
    url: "http://www.apache.org/liscences/LIS"
  servers:
    -url: http://localhost:3500
    description: THis is my local Server
    -url: http://example.com/
    description: This is my original URL
schemes: [https,http]
...
paths:
...
  /api/user/users/{id}:
    get:
      summary: Returns a user by ID.
      description: Returns the Specific User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type : integer
            format: int64
      responses: 
        '200':
          description: OK
          schema:
            type: object
            example: {name: syed}

#Post json/body & Body required
  # /create:
  #   post:
  #     summary: Adds a new user
  #     produces:
  #       - application/json
  #     consumes:
  #       - application/json
  #     parameters:
  #       - in: body
  #         name: body
  #         schema:
  #           type: object
  #           required:
  #             - name
  #             - id
  #           properties:
  #             id:
  #               type: integer
  #             name:
  #               type: string
  #     responses:
  #       '200':
  #         description: OK


#Defining Body Reference
...